Moved manipulation_history
